refactor: move session ID retrieval to repository layer
