+++ domain/src/main/java/com/amsterdam/domain/repository/UserListRepository.kt
@@ -1,13 +1,12 @@
 package com.amsterdam.domain.repository
 
-import com.amsterdam.entity.UserList
 import com.amsterdam.entity.Movie
+import com.amsterdam.entity.UserList
 
 interface UserListRepository {
     suspend fun getUserLists(
         accountId: Int = 0,
         page: Int = 1,
-        sessionId: String
     ): List<UserList>
 
     suspend fun getMoviesFromList(listId: Long, page: Int): List<Movie>
